+++ frontend/src/views/admin/CreateCompetitionPage.tsx
@@ -19,8 +19,8 @@ import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
 import { Switch } from "@/components/ui/switch";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { TimeInput } from "@/components/ui/TimeInput";
-import DatePicker from "@/components/ui/DatePicker";
+import { TimeInput } from "@/helpers/TimeInput";
+import DatePicker from "@/helpers/DatePicker";
 import { toast } from "sonner";
 import { createCompetition } from "@/api/CompetitionAPI";
 import { logFrontend } from "@/api/LoggerAPI";
